Convert fujiDeviceID_t to `enum class`

+++ lib/bus/rs232/FujiBusPacket.cpp
@@ -232,7 +232,7 @@ ByteBuffer FujiBusPacket::serialize() const
 {
     fujibus_header hdr, *hptr;
 
-    hdr.device = _device;
+    hdr.device = static_cast<uint8_t>(_device);
     hdr.command = static_cast<uint8_t>(_command);
     hdr.length = sizeof(hdr);
     hdr.checksum = 0;
